shared handles migrated to own header, made module entries
