+++ snippets/ObjectFree.hpp
@@ -41,36 +41,3 @@ private:
   Optional<const AllocationCallbacks> m_allocationCallbacks = nullptr;
   Dispatch const *                    m_dispatch            = nullptr;
 };
-
-//================================================================================================
-
-template <typename HandleType>
-class ObjectFreeShared
-{
-public:
-  using ParentType = parent_of_t<HandleType>;
-
-  template <class Dispatcher>
-  using destroy_pfn_t = void ( ParentType::* )( HandleType, const AllocationCallbacks *, const Dispatcher & ) const;
-
-  template <class Dispatcher = VULKAN_HPP_DEFAULT_DISPATCHER_TYPE>
-  ObjectFreeShared( Optional<const AllocationCallbacks> allocationCallbacks VULKAN_HPP_DEFAULT_ARGUMENT_NULLPTR_ASSIGNMENT,
-                    const Dispatcher & dispatch                             VULKAN_HPP_DEFAULT_DISPATCHER_ASSIGNMENT )
-    : m_destroy( reinterpret_cast<decltype( m_destroy )>( static_cast<destroy_pfn_t<Dispatcher>>( &ParentType::free ) ) )
-    , m_dispatch( &dispatch )
-    , m_allocationCallbacks( allocationCallbacks )
-  {
-  }
-
-public:
-  void destroy( ParentType parent, HandleType handle ) const VULKAN_HPP_NOEXCEPT
-  {
-    VULKAN_HPP_ASSERT( m_destroy && m_dispatch );
-    ( parent.*m_destroy )( handle, m_allocationCallbacks, *m_dispatch );
-  }
-
-private:
-  destroy_pfn_t<DispatchLoaderBase>   m_destroy             = nullptr;
-  const DispatchLoaderBase *          m_dispatch            = nullptr;
-  Optional<const AllocationCallbacks> m_allocationCallbacks = nullptr;
-};
