+++ vulkan/vulkan.hpp
@@ -1346,251 +1346,6 @@ namespace VULKAN_HPP_NAMESPACE
     lhs.swap( rhs );
   }
 #  endif
-
-#  if !defined( VULKAN_HPP_NO_SMART_HANDLE )
-
-  template <typename HandleType>
-  class SharedHandleTraits;
-
-  class NoParent;
-
-  template <typename HandleType, typename = void>
-  struct HasParentType : std::false_type
-  {
-  };
-
-  template <typename HandleType>
-  struct HasParentType<HandleType, decltype( (void)typename HandleType::ParentType() )> : std::true_type
-  {
-  };
-
-  template <typename HandleType, typename Enable = void>
-  struct GetParentType
-  {
-    using type = NoParent;
-  };
-
-  template <typename HandleType>
-  struct GetParentType<HandleType, typename std::enable_if<HasParentType<HandleType>::value>::type>
-  {
-    using type = typename HandleType::ParentType;
-  };
-
-  template <class HandleType>
-  using parent_of_t = typename GetParentType<HandleType>::type;
-
-  template <class HandleType>
-  VULKAN_HPP_CONSTEXPR_INLINE bool has_parent = !std::is_same<parent_of_t<HandleType>, NoParent>::value;
-
-  //=====================================================================================================================
-
-  template <typename HandleType>
-  class SharedHandle;
-
-  template <typename ParentType, typename Deleter>
-  struct SharedHeader
-  {
-    SharedHeader( SharedHandle<ParentType> parent, Deleter deleter = Deleter() ) VULKAN_HPP_NOEXCEPT
-      : parent( std::move( parent ) )
-      , deleter( std::move( deleter ) )
-    {
-    }
-
-    SharedHandle<ParentType> parent{};
-    Deleter                  deleter;
-  };
-
-  template <typename Deleter>
-  struct SharedHeader<NoParent, Deleter>
-  {
-    SharedHeader( Deleter deleter = Deleter() ) VULKAN_HPP_NOEXCEPT : deleter( std::move( deleter ) ) {}
-
-    Deleter deleter;
-  };
-
-  //=====================================================================================================================
-
-  template <typename HeaderType>
-  class ReferenceCounter
-  {
-  public:
-    template <typename... Args>
-    ReferenceCounter( Args &&... control_args ) : m_header( std::forward<Args>( control_args )... ){};
-    ReferenceCounter( const ReferenceCounter & )             = delete;
-    ReferenceCounter & operator=( const ReferenceCounter & ) = delete;
-
-  public:
-    size_t addRef() VULKAN_HPP_NOEXCEPT
-    {
-      return m_ref_cnt.fetch_add( 1, std::memory_order_relaxed );
-    }
-
-    size_t release() VULKAN_HPP_NOEXCEPT
-    {
-      return m_ref_cnt.fetch_sub( 1, std::memory_order_acq_rel );
-    }
-
-  public:
-    std::atomic_size_t m_ref_cnt{ 1 };
-    HeaderType         m_header{};
-  };
-
-  //=====================================================================================================================
-
-  template <typename HandleType, typename HeaderType, typename ForwardType = SharedHandle<HandleType>>
-  class SharedHandleBase
-  {
-  public:
-    SharedHandleBase() = default;
-
-    template <typename... Args>
-    SharedHandleBase( HandleType handle, Args &&... control_args )
-      : m_control( new ReferenceCounter<HeaderType>( std::forward<Args>( control_args )... ) ), m_handle( handle )
-    {
-    }
-
-    SharedHandleBase( const SharedHandleBase & o ) VULKAN_HPP_NOEXCEPT
-    {
-      o.addRef();
-      m_handle  = o.m_handle;
-      m_control = o.m_control;
-    }
-
-    SharedHandleBase( SharedHandleBase && o ) VULKAN_HPP_NOEXCEPT
-      : m_handle( o.m_handle )
-      , m_control( o.m_control )
-    {
-      o.m_handle  = nullptr;
-      o.m_control = nullptr;
-    }
-
-    SharedHandleBase & operator=( const SharedHandleBase & o ) VULKAN_HPP_NOEXCEPT
-    {
-      SharedHandleBase( o ).swap( *this );
-      return *this;
-    }
-
-    SharedHandleBase & operator=( SharedHandleBase && o ) VULKAN_HPP_NOEXCEPT
-    {
-      SharedHandleBase( std::move( o ) ).swap( *this );
-      return *this;
-    }
-
-    ~SharedHandleBase()
-    {
-      // only this function owns the last reference to the control block
-      // the same principle is used in the default deleter of std::shared_ptr
-      if ( m_control && ( m_control->release() == 1 ) )
-      {
-        ForwardType::internalDestroy( getHeader(), m_handle );
-        delete m_control;
-      }
-    }
-
-  public:
-    HandleType get() const VULKAN_HPP_NOEXCEPT
-    {
-      return m_handle;
-    }
-
-    explicit operator bool() const VULKAN_HPP_NOEXCEPT
-    {
-      return bool( m_handle );
-    }
-
-    const HandleType * operator->() const VULKAN_HPP_NOEXCEPT
-    {
-      return &m_handle;
-    }
-
-    HandleType * operator->() VULKAN_HPP_NOEXCEPT
-    {
-      return &m_handle;
-    }
-
-    void reset() VULKAN_HPP_NOEXCEPT
-    {
-      SharedHandleBase().swap( *this );
-    }
-
-    void swap( SharedHandleBase & o ) VULKAN_HPP_NOEXCEPT
-    {
-      std::swap( m_handle, o.m_handle );
-      std::swap( m_control, o.m_control );
-    }
-
-    template <typename T = HandleType>
-    typename std::enable_if<has_parent<T>, const SharedHandle<parent_of_t<HandleType>> &>::type getParent() const VULKAN_HPP_NOEXCEPT
-    {
-      return getHeader().parent;
-    }
-
-  protected:
-    template <typename T = HandleType>
-    static typename std::enable_if<!has_parent<T>, void>::type internalDestroy( const HeaderType & control, HandleType handle ) VULKAN_HPP_NOEXCEPT
-    {
-      control.deleter.destroy( handle );
-    }
-
-    template <typename T = HandleType>
-    static typename std::enable_if<has_parent<T>, void>::type internalDestroy( const HeaderType & control, HandleType handle ) VULKAN_HPP_NOEXCEPT
-    {
-      control.deleter.destroy( control.parent.get(), handle );
-    }
-
-    const HeaderType & getHeader() const VULKAN_HPP_NOEXCEPT
-    {
-      return m_control->m_header;
-    }
-
-  private:
-    void addRef() const VULKAN_HPP_NOEXCEPT
-    {
-      if ( m_control )
-        m_control->addRef();
-    }
-
-  protected:
-    ReferenceCounter<HeaderType> * m_control = nullptr;
-    HandleType                     m_handle{};
-  };
-
-  template <typename HandleType>
-  class SharedHandle : public SharedHandleBase<HandleType, SharedHeader<parent_of_t<HandleType>, typename SharedHandleTraits<HandleType>::deleter>>
-  {
-  private:
-    using BaseType    = SharedHandleBase<HandleType, SharedHeader<parent_of_t<HandleType>, typename SharedHandleTraits<HandleType>::deleter>>;
-    using DeleterType = typename SharedHandleTraits<HandleType>::deleter;
-    friend BaseType;
-
-  public:
-    using element_type = HandleType;
-
-    SharedHandle() = default;
-
-    template <typename T = HandleType, typename = typename std::enable_if<has_parent<T>>::type>
-    explicit SharedHandle( HandleType handle, SharedHandle<parent_of_t<HandleType>> parent, DeleterType deleter = DeleterType() ) VULKAN_HPP_NOEXCEPT
-      : BaseType( handle, std::move( parent ), std::move( deleter ) )
-    {
-    }
-
-    template <typename T = HandleType, typename = typename std::enable_if<!has_parent<T>>::type>
-    explicit SharedHandle( HandleType handle, DeleterType deleter = DeleterType() ) VULKAN_HPP_NOEXCEPT : BaseType( handle, std::move( deleter ) )
-    {
-    }
-
-  protected:
-    using BaseType::internalDestroy;
-  };
-
-  template <typename SharedType>
-  VULKAN_HPP_INLINE std::vector<typename SharedType::element_type> sharedToRaw( std::vector<SharedType> const & handles )
-  {
-    std::vector<typename SharedType::element_type> newBuffer( handles.size() );
-    std::transform( handles.begin(), handles.end(), newBuffer.begin(), []( SharedType const & handle ) { return handle.get(); } );
-    return newBuffer;
-  }
-#  endif
 #endif  // VULKAN_HPP_DISABLE_ENHANCED_MODE
 
   class DispatchLoaderBase
@@ -6406,52 +6161,6 @@ namespace VULKAN_HPP_NAMESPACE
     Dispatch const *                    m_dispatch            = nullptr;
   };
 
-  //================================================================================================
-
-  template <typename HandleType>
-  class ObjectDestroyShared
-  {
-  public:
-    using ParentType = parent_of_t<HandleType>;
-
-    template <class Dispatcher>
-    using destroy_pfn_t = typename std::conditional<has_parent<HandleType>,
-
-                                                    void ( ParentType::* )( HandleType, const AllocationCallbacks *, const Dispatcher & ) const,
-
-                                                    void ( HandleType::* )( const AllocationCallbacks *, const Dispatcher & ) const>::type;
-    using selector_t    = typename std::conditional<has_parent<HandleType>, ParentType, HandleType>::type;
-
-    template <typename Dispatcher = VULKAN_HPP_DEFAULT_DISPATCHER_TYPE>
-    ObjectDestroyShared( Optional<const AllocationCallbacks> allocationCallbacks VULKAN_HPP_DEFAULT_ARGUMENT_NULLPTR_ASSIGNMENT,
-                         const Dispatcher & dispatch                             VULKAN_HPP_DEFAULT_DISPATCHER_ASSIGNMENT )
-      : m_destroy( reinterpret_cast<decltype( m_destroy )>( static_cast<destroy_pfn_t<Dispatcher>>( &selector_t::destroy ) ) )
-      , m_dispatch( &dispatch )
-      , m_allocationCallbacks( allocationCallbacks )
-    {
-    }
-
-  public:
-    template <typename T = HandleType>
-    typename std::enable_if<has_parent<T>, void>::type destroy( ParentType parent, HandleType handle ) const VULKAN_HPP_NOEXCEPT
-    {
-      VULKAN_HPP_ASSERT( m_destroy && m_dispatch );
-      ( parent.*m_destroy )( handle, m_allocationCallbacks, *m_dispatch );
-    }
-
-    template <typename T = HandleType>
-    typename std::enable_if<!has_parent<T>, void>::type destroy( HandleType handle ) const VULKAN_HPP_NOEXCEPT
-    {
-      VULKAN_HPP_ASSERT( m_destroy && m_dispatch );
-      ( handle.*m_destroy )( m_allocationCallbacks, *m_dispatch );
-    }
-
-  private:
-    destroy_pfn_t<DispatchLoaderBase>   m_destroy             = nullptr;
-    const DispatchLoaderBase *          m_dispatch            = nullptr;
-    Optional<const AllocationCallbacks> m_allocationCallbacks = nullptr;
-  };
-
   template <typename OwnerType, typename Dispatch>
   class ObjectFree
   {
@@ -6496,39 +6205,6 @@ namespace VULKAN_HPP_NAMESPACE
     Dispatch const *                    m_dispatch            = nullptr;
   };
 
-  //================================================================================================
-
-  template <typename HandleType>
-  class ObjectFreeShared
-  {
-  public:
-    using ParentType = parent_of_t<HandleType>;
-
-    template <class Dispatcher>
-    using destroy_pfn_t = void ( ParentType::* )( HandleType, const AllocationCallbacks *, const Dispatcher & ) const;
-
-    template <class Dispatcher = VULKAN_HPP_DEFAULT_DISPATCHER_TYPE>
-    ObjectFreeShared( Optional<const AllocationCallbacks> allocationCallbacks VULKAN_HPP_DEFAULT_ARGUMENT_NULLPTR_ASSIGNMENT,
-                      const Dispatcher & dispatch                             VULKAN_HPP_DEFAULT_DISPATCHER_ASSIGNMENT )
-      : m_destroy( reinterpret_cast<decltype( m_destroy )>( static_cast<destroy_pfn_t<Dispatcher>>( &ParentType::free ) ) )
-      , m_dispatch( &dispatch )
-      , m_allocationCallbacks( allocationCallbacks )
-    {
-    }
-
-  public:
-    void destroy( ParentType parent, HandleType handle ) const VULKAN_HPP_NOEXCEPT
-    {
-      VULKAN_HPP_ASSERT( m_destroy && m_dispatch );
-      ( parent.*m_destroy )( handle, m_allocationCallbacks, *m_dispatch );
-    }
-
-  private:
-    destroy_pfn_t<DispatchLoaderBase>   m_destroy             = nullptr;
-    const DispatchLoaderBase *          m_dispatch            = nullptr;
-    Optional<const AllocationCallbacks> m_allocationCallbacks = nullptr;
-  };
-
   template <typename OwnerType, typename Dispatch>
   class ObjectRelease
   {
@@ -6562,36 +6238,6 @@ namespace VULKAN_HPP_NAMESPACE
   private:
     OwnerType        m_owner    = {};
     Dispatch const * m_dispatch = nullptr;
-  };
-
-  //================================================================================================
-
-  template <typename HandleType>
-  class ObjectReleaseShared
-  {
-  public:
-    using ParentType = parent_of_t<HandleType>;
-
-    template <class Dispatcher>
-    using destroy_pfn_t = void ( ParentType::* )( HandleType, const Dispatcher & ) const;
-
-    template <class Dispatcher = VULKAN_HPP_DEFAULT_DISPATCHER_TYPE>
-    ObjectReleaseShared( Optional<const AllocationCallbacks> allocationCallbacks VULKAN_HPP_DEFAULT_ARGUMENT_NULLPTR_ASSIGNMENT,
-                         const Dispatcher & dispatch                             VULKAN_HPP_DEFAULT_DISPATCHER_ASSIGNMENT )
-      : m_destroy( reinterpret_cast<decltype( m_destroy )>( static_cast<destroy_pfn_t<Dispatcher>>( &ParentType::release ) ) ), m_dispatch( &dispatch )
-    {
-    }
-
-  public:
-    void destroy( ParentType parent, HandleType handle ) const VULKAN_HPP_NOEXCEPT
-    {
-      VULKAN_HPP_ASSERT( m_destroy && m_dispatch );
-      ( parent.*m_destroy )( handle, *m_dispatch );
-    }
-
-  private:
-    destroy_pfn_t<DispatchLoaderBase> m_destroy  = nullptr;
-    const DispatchLoaderBase *        m_dispatch = nullptr;
   };
 
   template <typename OwnerType, typename PoolType, typename Dispatch>
@@ -6635,40 +6281,6 @@ namespace VULKAN_HPP_NAMESPACE
     Dispatch const * m_dispatch = nullptr;
   };
 
-  template <typename HandleType, typename PoolType>
-  class PoolFreeShared
-  {
-  public:
-    using ParentType = parent_of_t<HandleType>;
-
-    template <class Dispatcher>
-    using ret_t = decltype( std::declval<ParentType>().free( PoolType(), 0u, nullptr, Dispatcher() ) );
-
-    template <class Dispatcher>
-    using destroy_pfn_t = ret_t<Dispatcher> ( ParentType::* )( PoolType, uint32_t, const HandleType *, const Dispatcher & ) const;
-
-    PoolFreeShared() = default;
-
-    template <class Dispatcher = VULKAN_HPP_DEFAULT_DISPATCHER_TYPE>
-    PoolFreeShared( SharedHandle<PoolType> pool, const Dispatcher & dispatch VULKAN_HPP_DEFAULT_DISPATCHER_ASSIGNMENT )
-      : m_destroy( reinterpret_cast<decltype( m_destroy )>( static_cast<destroy_pfn_t<Dispatcher>>( &ParentType::free ) ) )
-      , m_pool( std::move( pool ) )
-      , m_dispatch( &dispatch )
-    {
-    }
-
-  public:
-    void destroy( ParentType parent, HandleType handle ) const VULKAN_HPP_NOEXCEPT
-    {
-      VULKAN_HPP_ASSERT( m_destroy && m_dispatch );
-      ( parent.*m_destroy )( m_pool.get(), 1u, &handle, *m_dispatch );
-    }
-
-  private:
-    destroy_pfn_t<DispatchLoaderBase> m_destroy  = nullptr;
-    const DispatchLoaderBase *        m_dispatch = nullptr;
-    SharedHandle<PoolType>            m_pool{};
-  };
 #endif  // !VULKAN_HPP_NO_SMART_HANDLE
 
   //==================
@@ -7425,197 +7037,12 @@ namespace VULKAN_HPP_NAMESPACE
 #include <vulkan/vulkan_handles.hpp>
 #include <vulkan/vulkan_structs.hpp>
 #include <vulkan/vulkan_funcs.hpp>
+#include <vulkan/vulkan_shared.hpp>
 
 // clang-format on
 
 namespace VULKAN_HPP_NAMESPACE
 {
-#if !defined( VULKAN_HPP_NO_SMART_HANDLE )
-
-  enum class SwapchainOwns
-  {
-    no,
-    yes,
-  };
-
-  struct ImageHeader : SharedHeader<parent_of_t<VULKAN_HPP_NAMESPACE::Image>, typename SharedHandleTraits<VULKAN_HPP_NAMESPACE::Image>::deleter>
-  {
-    ImageHeader(
-      SharedHandle<parent_of_t<VULKAN_HPP_NAMESPACE::Image>>            parent,
-      typename SharedHandleTraits<VULKAN_HPP_NAMESPACE::Image>::deleter deleter        = typename SharedHandleTraits<VULKAN_HPP_NAMESPACE::Image>::deleter(),
-      SwapchainOwns                                                     swapchainOwned = SwapchainOwns::no ) VULKAN_HPP_NOEXCEPT
-      : SharedHeader<parent_of_t<VULKAN_HPP_NAMESPACE::Image>, typename SharedHandleTraits<VULKAN_HPP_NAMESPACE::Image>::deleter>( std::move( parent ),
-                                                                                                                                   std::move( deleter ) )
-      , swapchainOwned( swapchainOwned )
-    {
-    }
-
-    SwapchainOwns swapchainOwned = SwapchainOwns::no;
-  };
-
-  template <>
-  class SharedHandle<VULKAN_HPP_NAMESPACE::Image> : public SharedHandleBase<VULKAN_HPP_NAMESPACE::Image, ImageHeader>
-  {
-    using BaseType    = SharedHandleBase<VULKAN_HPP_NAMESPACE::Image, ImageHeader>;
-    using DeleterType = typename SharedHandleTraits<VULKAN_HPP_NAMESPACE::Image>::deleter;
-    friend BaseType;
-
-  public:
-    using element_type = VULKAN_HPP_NAMESPACE::Image;
-
-  public:
-    SharedHandle() = default;
-
-    explicit SharedHandle( VULKAN_HPP_NAMESPACE::Image                            handle,
-                           SharedHandle<parent_of_t<VULKAN_HPP_NAMESPACE::Image>> parent,
-                           SwapchainOwns                                          swapchain_owned = SwapchainOwns::no,
-                           DeleterType                                            deleter         = DeleterType() ) VULKAN_HPP_NOEXCEPT
-      : BaseType( handle, std::move( parent ), std::move( deleter ), swapchain_owned )
-    {
-    }
-
-  protected:
-    static void internalDestroy( const ImageHeader & control, VULKAN_HPP_NAMESPACE::Image handle ) VULKAN_HPP_NOEXCEPT
-    {
-      if ( control.swapchainOwned == SwapchainOwns::no )
-      {
-        control.deleter.destroy( control.parent.get(), handle );
-      }
-    }
-  };
-
-  struct SwapchainHeader
-  {
-    SwapchainHeader( SharedHandle<VULKAN_HPP_NAMESPACE::SurfaceKHR>                           surface,
-                     SharedHandle<parent_of_t<VULKAN_HPP_NAMESPACE::SwapchainKHR>>            parent,
-                     typename SharedHandleTraits<VULKAN_HPP_NAMESPACE::SwapchainKHR>::deleter deleter =
-                       typename SharedHandleTraits<VULKAN_HPP_NAMESPACE::SwapchainKHR>::deleter() ) VULKAN_HPP_NOEXCEPT
-      : surface( std::move( surface ) )
-      , parent( std::move( parent ) )
-      , deleter( std::move( deleter ) )
-    {
-    }
-
-    SharedHandle<VULKAN_HPP_NAMESPACE::SurfaceKHR>                           surface{};
-    SharedHandle<parent_of_t<VULKAN_HPP_NAMESPACE::SwapchainKHR>>            parent{};
-    typename SharedHandleTraits<VULKAN_HPP_NAMESPACE::SwapchainKHR>::deleter deleter{};
-  };
-
-  template <>
-  class SharedHandle<VULKAN_HPP_NAMESPACE::SwapchainKHR> : public SharedHandleBase<VULKAN_HPP_NAMESPACE::SwapchainKHR, SwapchainHeader>
-  {
-    using BaseType    = SharedHandleBase<VULKAN_HPP_NAMESPACE::SwapchainKHR, SwapchainHeader>;
-    using DeleterType = typename SharedHandleTraits<VULKAN_HPP_NAMESPACE::SwapchainKHR>::deleter;
-    friend BaseType;
-
-  public:
-    using element_type = VULKAN_HPP_NAMESPACE::SwapchainKHR;
-
-  public:
-    SharedHandle() = default;
-
-    explicit SharedHandle( VULKAN_HPP_NAMESPACE::SwapchainKHR                            handle,
-                           SharedHandle<parent_of_t<VULKAN_HPP_NAMESPACE::SwapchainKHR>> parent,
-                           SharedHandle<VULKAN_HPP_NAMESPACE::SurfaceKHR>                surface,
-                           DeleterType                                                   deleter = DeleterType() ) VULKAN_HPP_NOEXCEPT
-      : BaseType( handle, std::move( surface ), std::move( parent ), std::move( deleter ) )
-    {
-    }
-
-  public:
-    const SharedHandle<VULKAN_HPP_NAMESPACE::SurfaceKHR> & getSurface() const VULKAN_HPP_NOEXCEPT
-    {
-      return getHeader().surface;
-    }
-
-  protected:
-    using BaseType::internalDestroy;
-  };
-
-  template <typename HandleType, typename ParentType>
-  class SharedHandleBaseNoDestroy : public SharedHandleBase<HandleType, ParentType>
-  {
-    using BaseType = SharedHandleBase<HandleType, ParentType>;
-
-  public:
-    using BaseType::SharedHandleBase;
-
-    const ParentType & getParent() const VULKAN_HPP_NOEXCEPT
-    {
-      return BaseType::getHeader();
-    }
-
-  protected:
-    static void internalDestroy( const ParentType &, HandleType ) VULKAN_HPP_NOEXCEPT {}
-  };
-
-  //=== VK_VERSION_1_0 ===
-
-  template <>
-  class SharedHandle<PhysicalDevice> : public SharedHandleBaseNoDestroy<PhysicalDevice, SharedInstance>
-  {
-    friend SharedHandleBase<PhysicalDevice, SharedInstance>;
-
-  public:
-    using element_type = PhysicalDevice;
-    using SharedHandleBaseNoDestroy<PhysicalDevice, SharedInstance>::SharedHandleBaseNoDestroy;
-  };
-
-  using SharedPhysicalDevice = SharedHandle<PhysicalDevice>;
-
-  template <>
-  class SharedHandle<Queue> : public SharedHandleBaseNoDestroy<Queue, SharedDevice>
-  {
-    friend SharedHandleBase<Queue, SharedDevice>;
-
-  public:
-    using element_type = Queue;
-    using SharedHandleBaseNoDestroy<Queue, SharedDevice>::SharedHandleBaseNoDestroy;
-  };
-
-  using SharedQueue = SharedHandle<Queue>;
-
-  //=== VK_KHR_display ===
-
-  template <>
-  class SharedHandle<DisplayKHR> : public SharedHandleBaseNoDestroy<DisplayKHR, SharedPhysicalDevice>
-  {
-    friend SharedHandleBase<DisplayKHR, SharedPhysicalDevice>;
-
-  public:
-    using element_type = DisplayKHR;
-    using SharedHandleBaseNoDestroy<DisplayKHR, SharedPhysicalDevice>::SharedHandleBaseNoDestroy;
-  };
-
-  using SharedDisplayKHR = SharedHandle<DisplayKHR>;
-
-  template <>
-  class SharedHandle<DisplayModeKHR> : public SharedHandleBaseNoDestroy<DisplayModeKHR, SharedDisplayKHR>
-  {
-    friend SharedHandleBase<DisplayModeKHR, SharedDisplayKHR>;
-
-  public:
-    using element_type = DisplayModeKHR;
-    using SharedHandleBaseNoDestroy<DisplayModeKHR, SharedDisplayKHR>::SharedHandleBaseNoDestroy;
-  };
-
-  using SharedDisplayModeKHR = SharedHandle<DisplayModeKHR>;
-
-  //=== VK_INTEL_performance_query ===
-
-  template <>
-  class SharedHandle<PerformanceConfigurationINTEL> : public SharedHandleBaseNoDestroy<PerformanceConfigurationINTEL, SharedDevice>
-  {
-    friend SharedHandleBase<PerformanceConfigurationINTEL, SharedDevice>;
-
-  public:
-    using element_type = PerformanceConfigurationINTEL;
-    using SharedHandleBaseNoDestroy<PerformanceConfigurationINTEL, SharedDevice>::SharedHandleBaseNoDestroy;
-  };
-
-  using SharedPerformanceConfigurationINTEL = SharedHandle<PerformanceConfigurationINTEL>;
-#endif  // !VULKAN_HPP_NO_SMART_HANDLE
-
 #if !defined( VULKAN_HPP_DISABLE_ENHANCED_MODE )
 
   //=======================
