+++ vulkan/vulkan_handles.hpp
@@ -2153,422 +2153,6 @@ namespace VULKAN_HPP_NAMESPACE
   using UniqueShaderEXT = UniqueHandle<ShaderEXT, VULKAN_HPP_DEFAULT_DISPATCHER_TYPE>;
 #endif /*VULKAN_HPP_NO_SMART_HANDLE*/
 
-#ifndef VULKAN_HPP_NO_SMART_HANDLE
-  //======================
-  //=== SHARED HANDLEs ===
-  //======================
-
-  //=== VK_VERSION_1_0 ===
-  template <>
-  class SharedHandleTraits<Instance>
-  {
-  public:
-    using deleter = ObjectDestroyShared<Instance>;
-  };
-
-  using SharedInstance = SharedHandle<Instance>;
-
-  template <>
-  class SharedHandleTraits<Device>
-  {
-  public:
-    using deleter = ObjectDestroyShared<Device>;
-  };
-
-  using SharedDevice = SharedHandle<Device>;
-
-  template <>
-  class SharedHandleTraits<DeviceMemory>
-  {
-  public:
-    using deleter = ObjectFreeShared<DeviceMemory>;
-  };
-
-  using SharedDeviceMemory = SharedHandle<DeviceMemory>;
-
-  template <>
-  class SharedHandleTraits<Fence>
-  {
-  public:
-    using deleter = ObjectDestroyShared<Fence>;
-  };
-
-  using SharedFence = SharedHandle<Fence>;
-
-  template <>
-  class SharedHandleTraits<Semaphore>
-  {
-  public:
-    using deleter = ObjectDestroyShared<Semaphore>;
-  };
-
-  using SharedSemaphore = SharedHandle<Semaphore>;
-
-  template <>
-  class SharedHandleTraits<Event>
-  {
-  public:
-    using deleter = ObjectDestroyShared<Event>;
-  };
-
-  using SharedEvent = SharedHandle<Event>;
-
-  template <>
-  class SharedHandleTraits<QueryPool>
-  {
-  public:
-    using deleter = ObjectDestroyShared<QueryPool>;
-  };
-
-  using SharedQueryPool = SharedHandle<QueryPool>;
-
-  template <>
-  class SharedHandleTraits<Buffer>
-  {
-  public:
-    using deleter = ObjectDestroyShared<Buffer>;
-  };
-
-  using SharedBuffer = SharedHandle<Buffer>;
-
-  template <>
-  class SharedHandleTraits<BufferView>
-  {
-  public:
-    using deleter = ObjectDestroyShared<BufferView>;
-  };
-
-  using SharedBufferView = SharedHandle<BufferView>;
-
-  template <>
-  class SharedHandleTraits<Image>
-  {
-  public:
-    using deleter = ObjectDestroyShared<Image>;
-  };
-
-  using SharedImage = SharedHandle<Image>;
-
-  template <>
-  class SharedHandleTraits<ImageView>
-  {
-  public:
-    using deleter = ObjectDestroyShared<ImageView>;
-  };
-
-  using SharedImageView = SharedHandle<ImageView>;
-
-  template <>
-  class SharedHandleTraits<ShaderModule>
-  {
-  public:
-    using deleter = ObjectDestroyShared<ShaderModule>;
-  };
-
-  using SharedShaderModule = SharedHandle<ShaderModule>;
-
-  template <>
-  class SharedHandleTraits<PipelineCache>
-  {
-  public:
-    using deleter = ObjectDestroyShared<PipelineCache>;
-  };
-
-  using SharedPipelineCache = SharedHandle<PipelineCache>;
-
-  template <>
-  class SharedHandleTraits<Pipeline>
-  {
-  public:
-    using deleter = ObjectDestroyShared<Pipeline>;
-  };
-
-  using SharedPipeline = SharedHandle<Pipeline>;
-
-  template <>
-  class SharedHandleTraits<PipelineLayout>
-  {
-  public:
-    using deleter = ObjectDestroyShared<PipelineLayout>;
-  };
-
-  using SharedPipelineLayout = SharedHandle<PipelineLayout>;
-
-  template <>
-  class SharedHandleTraits<Sampler>
-  {
-  public:
-    using deleter = ObjectDestroyShared<Sampler>;
-  };
-
-  using SharedSampler = SharedHandle<Sampler>;
-
-  template <>
-  class SharedHandleTraits<DescriptorPool>
-  {
-  public:
-    using deleter = ObjectDestroyShared<DescriptorPool>;
-  };
-
-  using SharedDescriptorPool = SharedHandle<DescriptorPool>;
-
-  template <>
-  class SharedHandleTraits<DescriptorSet>
-  {
-  public:
-    using deleter = PoolFreeShared<DescriptorSet, DescriptorPool>;
-  };
-
-  using SharedDescriptorSet = SharedHandle<DescriptorSet>;
-
-  template <>
-  class SharedHandleTraits<DescriptorSetLayout>
-  {
-  public:
-    using deleter = ObjectDestroyShared<DescriptorSetLayout>;
-  };
-
-  using SharedDescriptorSetLayout = SharedHandle<DescriptorSetLayout>;
-
-  template <>
-  class SharedHandleTraits<Framebuffer>
-  {
-  public:
-    using deleter = ObjectDestroyShared<Framebuffer>;
-  };
-
-  using SharedFramebuffer = SharedHandle<Framebuffer>;
-
-  template <>
-  class SharedHandleTraits<RenderPass>
-  {
-  public:
-    using deleter = ObjectDestroyShared<RenderPass>;
-  };
-
-  using SharedRenderPass = SharedHandle<RenderPass>;
-
-  template <>
-  class SharedHandleTraits<CommandPool>
-  {
-  public:
-    using deleter = ObjectDestroyShared<CommandPool>;
-  };
-
-  using SharedCommandPool = SharedHandle<CommandPool>;
-
-  template <>
-  class SharedHandleTraits<CommandBuffer>
-  {
-  public:
-    using deleter = PoolFreeShared<CommandBuffer, CommandPool>;
-  };
-
-  using SharedCommandBuffer = SharedHandle<CommandBuffer>;
-
-  //=== VK_VERSION_1_1 ===
-  template <>
-  class SharedHandleTraits<SamplerYcbcrConversion>
-  {
-  public:
-    using deleter = ObjectDestroyShared<SamplerYcbcrConversion>;
-  };
-
-  using SharedSamplerYcbcrConversion    = SharedHandle<SamplerYcbcrConversion>;
-  using SharedSamplerYcbcrConversionKHR = SharedHandle<SamplerYcbcrConversion>;
-
-  template <>
-  class SharedHandleTraits<DescriptorUpdateTemplate>
-  {
-  public:
-    using deleter = ObjectDestroyShared<DescriptorUpdateTemplate>;
-  };
-
-  using SharedDescriptorUpdateTemplate    = SharedHandle<DescriptorUpdateTemplate>;
-  using SharedDescriptorUpdateTemplateKHR = SharedHandle<DescriptorUpdateTemplate>;
-
-  //=== VK_VERSION_1_3 ===
-  template <>
-  class SharedHandleTraits<PrivateDataSlot>
-  {
-  public:
-    using deleter = ObjectDestroyShared<PrivateDataSlot>;
-  };
-
-  using SharedPrivateDataSlot    = SharedHandle<PrivateDataSlot>;
-  using SharedPrivateDataSlotEXT = SharedHandle<PrivateDataSlot>;
-
-  //=== VK_KHR_surface ===
-  template <>
-  class SharedHandleTraits<SurfaceKHR>
-  {
-  public:
-    using deleter = ObjectDestroyShared<SurfaceKHR>;
-  };
-
-  using SharedSurfaceKHR = SharedHandle<SurfaceKHR>;
-
-  //=== VK_KHR_swapchain ===
-  template <>
-  class SharedHandleTraits<SwapchainKHR>
-  {
-  public:
-    using deleter = ObjectDestroyShared<SwapchainKHR>;
-  };
-
-  using SharedSwapchainKHR = SharedHandle<SwapchainKHR>;
-
-  //=== VK_EXT_debug_report ===
-  template <>
-  class SharedHandleTraits<DebugReportCallbackEXT>
-  {
-  public:
-    using deleter = ObjectDestroyShared<DebugReportCallbackEXT>;
-  };
-
-  using SharedDebugReportCallbackEXT = SharedHandle<DebugReportCallbackEXT>;
-
-  //=== VK_KHR_video_queue ===
-  template <>
-  class SharedHandleTraits<VideoSessionKHR>
-  {
-  public:
-    using deleter = ObjectDestroyShared<VideoSessionKHR>;
-  };
-
-  using SharedVideoSessionKHR = SharedHandle<VideoSessionKHR>;
-
-  template <>
-  class SharedHandleTraits<VideoSessionParametersKHR>
-  {
-  public:
-    using deleter = ObjectDestroyShared<VideoSessionParametersKHR>;
-  };
-
-  using SharedVideoSessionParametersKHR = SharedHandle<VideoSessionParametersKHR>;
-
-  //=== VK_NVX_binary_import ===
-  template <>
-  class SharedHandleTraits<CuModuleNVX>
-  {
-  public:
-    using deleter = ObjectDestroyShared<CuModuleNVX>;
-  };
-
-  using SharedCuModuleNVX = SharedHandle<CuModuleNVX>;
-
-  template <>
-  class SharedHandleTraits<CuFunctionNVX>
-  {
-  public:
-    using deleter = ObjectDestroyShared<CuFunctionNVX>;
-  };
-
-  using SharedCuFunctionNVX = SharedHandle<CuFunctionNVX>;
-
-  //=== VK_EXT_debug_utils ===
-  template <>
-  class SharedHandleTraits<DebugUtilsMessengerEXT>
-  {
-  public:
-    using deleter = ObjectDestroyShared<DebugUtilsMessengerEXT>;
-  };
-
-  using SharedDebugUtilsMessengerEXT = SharedHandle<DebugUtilsMessengerEXT>;
-
-  //=== VK_KHR_acceleration_structure ===
-  template <>
-  class SharedHandleTraits<AccelerationStructureKHR>
-  {
-  public:
-    using deleter = ObjectDestroyShared<AccelerationStructureKHR>;
-  };
-
-  using SharedAccelerationStructureKHR = SharedHandle<AccelerationStructureKHR>;
-
-  //=== VK_EXT_validation_cache ===
-  template <>
-  class SharedHandleTraits<ValidationCacheEXT>
-  {
-  public:
-    using deleter = ObjectDestroyShared<ValidationCacheEXT>;
-  };
-
-  using SharedValidationCacheEXT = SharedHandle<ValidationCacheEXT>;
-
-  //=== VK_NV_ray_tracing ===
-  template <>
-  class SharedHandleTraits<AccelerationStructureNV>
-  {
-  public:
-    using deleter = ObjectDestroyShared<AccelerationStructureNV>;
-  };
-
-  using SharedAccelerationStructureNV = SharedHandle<AccelerationStructureNV>;
-
-  //=== VK_KHR_deferred_host_operations ===
-  template <>
-  class SharedHandleTraits<DeferredOperationKHR>
-  {
-  public:
-    using deleter = ObjectDestroyShared<DeferredOperationKHR>;
-  };
-
-  using SharedDeferredOperationKHR = SharedHandle<DeferredOperationKHR>;
-
-  //=== VK_NV_device_generated_commands ===
-  template <>
-  class SharedHandleTraits<IndirectCommandsLayoutNV>
-  {
-  public:
-    using deleter = ObjectDestroyShared<IndirectCommandsLayoutNV>;
-  };
-
-  using SharedIndirectCommandsLayoutNV = SharedHandle<IndirectCommandsLayoutNV>;
-
-#  if defined( VK_USE_PLATFORM_FUCHSIA )
-  //=== VK_FUCHSIA_buffer_collection ===
-  template <>
-  class SharedHandleTraits<BufferCollectionFUCHSIA>
-  {
-  public:
-    using deleter = ObjectDestroyShared<BufferCollectionFUCHSIA>;
-  };
-
-  using SharedBufferCollectionFUCHSIA = SharedHandle<BufferCollectionFUCHSIA>;
-#  endif /*VK_USE_PLATFORM_FUCHSIA*/
-
-  //=== VK_EXT_opacity_micromap ===
-  template <>
-  class SharedHandleTraits<MicromapEXT>
-  {
-  public:
-    using deleter = ObjectDestroyShared<MicromapEXT>;
-  };
-
-  using SharedMicromapEXT = SharedHandle<MicromapEXT>;
-
-  //=== VK_NV_optical_flow ===
-  template <>
-  class SharedHandleTraits<OpticalFlowSessionNV>
-  {
-  public:
-    using deleter = ObjectDestroyShared<OpticalFlowSessionNV>;
-  };
-
-  using SharedOpticalFlowSessionNV = SharedHandle<OpticalFlowSessionNV>;
-
-  //=== VK_EXT_shader_object ===
-  template <>
-  class SharedHandleTraits<ShaderEXT>
-  {
-  public:
-    using deleter = ObjectDestroyShared<ShaderEXT>;
-  };
-
-  using SharedShaderEXT = SharedHandle<ShaderEXT>;
-#endif /*VULKAN_HPP_NO_SMART_HANDLE*/
-
   //===============
   //=== HANDLEs ===
   //===============
